+++ Options/SettingDefinitions/Quests.lua
@@ -26,6 +26,25 @@ class "SettingDefinitions.Quests" (function(_ENV)
     enablePOICheckBox:BindUISetting("quest.enablePOI")
     self.GeneralTabControls.enablePOICheckBox = enablePOICheckBox
 
+    local questNewRemovePoliciesText = Widgets.SettingsText.Acquire(false, self)
+    questNewRemovePoliciesText:SetID(30)
+    questNewRemovePoliciesText:SetText("A quest is no longer considered new when:")
+    self.GeneralTabControls.questNewRemovePoliciesText = questNewRemovePoliciesText
+
+    local questNewMaxAgeSlider = Widgets.SettingsSlider.Acquire(false, self)
+    questNewMaxAgeSlider:SetID(40)
+    questNewMaxAgeSlider:SetLabel("- it is older than x seconds")
+    questNewMaxAgeSlider:SetMinMaxValues(0, 32140800)
+    questNewMaxAgeSlider:BindSetting("questNewMaxAge")
+    self.GeneralTabControls.questNewMaxAgeSlider = questNewMaxAgeSlider
+    Style[questNewMaxAgeSlider].marginLeft = 20
+
+    local questNewRemoveOnProgressCheckBox = Widgets.SettingsCheckBox.Acquire(false, self)
+    questNewRemoveOnProgressCheckBox:SetID(50)
+    questNewRemoveOnProgressCheckBox:SetLabel("- it has received progress")
+    questNewRemoveOnProgressCheckBox:BindSetting("questNewRemoveOnProgress")
+    self.GeneralTabControls.questNewRemoveOnProgressCheckBox = questNewRemoveOnProgressCheckBox
+    Style[questNewRemoveOnProgressCheckBox].marginLeft = 20
     -- ---------------------------------------------------------------------------
     -- --- Background Section
     -- ---------------------------------------------------------------------------
@@ -486,8 +505,16 @@ class "SettingDefinitions.Quests" (function(_ENV)
     showTooltip:BindUISetting("quest.showTooltip")
     self.QuestTabControls.showTooltip = showTooltip
 
+    if IsRetail() then 
+      local showNewQuestIndicator =  Widgets.SettingsCheckBox.Acquire(false, self)
+      showNewQuestIndicator:SetID(70)
+      showNewQuestIndicator:SetLabel("Show New Quest Indicator")
+      showNewQuestIndicator:BindUISetting("quest.showNewQuestIndicator")
+      self.QuestTabControls.showNewQuestIndicator = showNewQuestIndicator
+    end
+
     local showRewardsInTooltip =  Widgets.SettingsCheckBox.Acquire(false, self)
-    showRewardsInTooltip:SetID(70)
+    showRewardsInTooltip:SetID(80)
     showRewardsInTooltip:SetLabel(("|cffff7f00(%s)|r %s"):format(L.EXPERIMENTAL, "Show Rewards in Tooltip"))
     showRewardsInTooltip:BindUISetting("quest.tooltip.showRewards")
     self.QuestTabControls.showRewardsInTooltip = showRewardsInTooltip
@@ -496,7 +523,6 @@ class "SettingDefinitions.Quests" (function(_ENV)
     headertabControl:SelectTab(1)
 
     self.QuestTabControls.headertabControl = headertabControl
-
     ---------------------------------------------------------------------------
     --- Objectives Section
     ---------------------------------------------------------------------------